Add RequiredActions component with finish signup and accept updated terms modals
